% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study.R
\name{case_study_summary}
\alias{case_study_summary}
\title{Summarize Case Studies}
\usage{
case_study_summary(rmd_directory = "vignettes")
}
\arguments{
\item{rmd_directory}{\code{character(1)} directory in which case study
vignette source (\verb{hca-.*\\\\.Rmd}) are located.}
}
\description{
\code{case_study_summary()} visits each case study in the
'vignettes' directory of the package, summarizing the study and
outcome in a tibble.
}
