% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_tsv.R
\name{csv_tsv_count_lines}
\alias{csv_tsv_count_lines}
\title{Summarize Tab- or Comma-Separated Files}
\usage{
csv_tsv_count_lines(.data, file_id = "fileId", file_path = "local.filePath")
}
\arguments{
\item{.data}{\code{tibble} containing a column with \code{file_id} and
\code{file_path} columns.}

\item{file_id}{\code{character(1)} a column name in \code{.data} containing
(unique) identifier for the files for which line counts are
requested.}

\item{file_path}{\code{character(1)} a column name in \code{.data} containing
file paths to the file for which lines will be counted.}
}
\value{
\code{csv_tsv_count_lines()} returns the original tibble
\code{.data} with an additional column \code{local.lines} containing the
number of lines in \code{file_path}.
}
\description{
\code{csv_tsv_count_lines()} counts the number of lines
in one or several comma- or tab-delimited files.
}
