% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtx.R
\name{mtx_count_matrix}
\alias{mtx_count_matrix}
\title{Summarize 'MatrixMart' files}
\usage{
mtx_count_matrix(
  .data,
  mtx_path,
  genes_path = NULL,
  barcodes_path = NULL,
  sep = "\\t"
)
}
\arguments{
\item{.data}{\code{tibble} containing a column with the value of the
\code{mtx_path} parameter and, if non-NULL, columns with the value
of parameters \code{genes_path} and \code{barcodes_path}.}

\item{mtx_path}{\code{character()} paths to MatrixMart files on disk.}

\item{genes_path}{\code{character()} (optional) paths to files
containing gene (row) labels (e.g., ENSEMBL gene
identifiers). The row labels must be in the first column of the
file; fields are separated by the parameter \code{sep}; there can be
no column headers in the file.}

\item{barcodes_path}{\code{character()} (optional) paths to files
containing cell (column) labels (e.g., barcodes). The column
labels must be in the first column of the file; fields are
separated by the parameter \code{sep}; there can be no column
headers in the file.}

\item{sep}{\code{character(1)}, default \code{"\\t"} (tab). Field separator in
\code{genes_path} and \code{barcodes_path} files.}
}
\value{
\code{mtx_count_matrix()} returns a \code{dgCMatrix} sparse matrix
with rows (genes) equal to the unique genes in all input files,
and columns (cells) equal to the total cell count in the
MatrixMart files. Samples (columns) are in the order present in
\code{.data}.
}
\description{
\code{mtx_count_matrix()} parses one or several MatrixMart
files with identical dimensions into a single \code{dgCMatrix}
object.
}
