% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_template.R
\name{use_template}
\alias{use_template}
\title{Templates For New Case Studies}
\usage{
use_template(project_id, case_study_path, overwrite = FALSE)
}
\arguments{
\item{case_study_path}{character(1) path to the directory in which
the case study will be created.}

\item{overwrite}{logical(1) overwrite a case study created by this
function with the same \code{project_id} and \code{case_study_path}?
Default: \code{FALSE}.}
}
\value{
\code{use_template()} returns the path to the rendered template
(in \code{case_study_path}) with approriate substitution of
variables.
}
\description{
\code{use_template()} creates a new \emph{R} markdown file as a
starting point for a case study. The file requires manual
editing to customize for the specific project.
}
\details{
\code{use_template()} queries the HCA data portal for the
project title associated with \code{project_id}, and uses this for
the title and vignette index entry of the rendered markdown
doeucment. The \code{project_id} is injected into the document, so
that references to \code{project_id} are defined. Rendering the
template also injects the current date, package name, and
package version.
}
\examples{
project_id <- "3c9d586e-bd26-4b46-8690-3faaa18ccf38"
case_study_path <- tempfile(); dir.create(case_study_path)
rmd <- use_template(project_id, case_study_path)
readLines(rmd) |>
    head() |>
    noquote()

}
